utils.c: Add mp_debug() and mp_verbosity
mp_verbosity replaces the "verbose" and "debug" variables from
many of the scripts, and mp_debug() provides a simple helper
that does the if(mp_verbosity > X) { printf() } for us.

Not that it makes much difference for the existing plugins,
although we save 10-20 lines per 2 plugins migrated, so that's
alright, I guess.

Mostly though, it's handy when we're writing new plugins,
which is a mission I'm about to embark upon.

Signed-off-by: Andreas Ericsson <ae@op5.se>